requirements.txt
Flask==3.0.0
Werkzeug==3.0.1
python-dotenv==1.0.0
requests==2.31.0
python-dateutil==2.8.2
bleach==6.1.0
Flask-Limiter==3.5.0
supabase==2.3.0
